+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Changed the location of working directory
 - Added Ctrl+A and Ctrl+X shortcuts for the Image/Tilemap Editor
 - Added copy and paste feature for the Sound/Music Editor
+- Added the example #12
 
 ## 1.6.9
 - Fixed the select tool of the resource editor